Modified starting up the test-server in Rakefile:
Driving the 'rackup'-command inside a separate thread now utilizes popen3
to analyze the server's output with a custom REGEXP before yielding. This
is necessary to prevent firing up tests before the server is ready to accept
any input.

The server's output can still be displayed on STDOUT by providing 'DEBUG=true'
to 'rake test' or 'rake test_server:start', though the mechanism is falls back
to wait for a given number of seconds before yielding